chore: refactor diff cluster command
