+++ internal/apis/kfd/v1alpha2/ekscluster/init.go
@@ -22,4 +22,10 @@ func init() {
 		"EKSCluster",
 		cluster.NewDeleterFactory[*ClusterDeleter, private.EksclusterKfdV1Alpha2](&ClusterDeleter{}),
 	)
+
+	cluster.RegisterSchemaSettings(
+		"kfd.sighup.io/v1alpha2",
+		"EKSCluster",
+		NewSchemaSettings(),
+	)
 }
