+++ internal/apis/kfd/v1alpha2/ekscluster/schema.go
@@ -8,7 +8,16 @@ import (
 	"fmt"
 
 	"github.com/sighupio/fury-distribution/pkg/apis/ekscluster/v1alpha2/private"
+	"github.com/sighupio/furyctl/internal/cluster"
 	yamlx "github.com/sighupio/furyctl/internal/x/yaml"
+)
+
+const (
+	InfrastructurePhaseSchemaPath = ".spec.infrastructure"
+	KubernetesPhaseSchemaPath     = ".spec.kubernetes"
+	DistributionPhaseSchemaPath   = ".spec.distribution"
+	PluginsPhaseSchemaPath        = ".spec.plugins"
+	AllPhaseSchemaPath            = ""
 )
 
 var ErrInvalidNodePoolSize = fmt.Errorf("invalid node pool size")
@@ -34,4 +43,32 @@ func (*ExtraSchemaValidator) Validate(confPath string) error {
 	}
 
 	return nil
+}
+
+func NewSchemaSettings() *SchemaSettings {
+	return &SchemaSettings{}
+}
+
+type SchemaSettings struct{}
+
+func (*SchemaSettings) SchemaPathForPhase(phase string) (string, error) {
+	switch phase {
+	case cluster.OperationPhaseInfrastructure:
+		return InfrastructurePhaseSchemaPath, nil
+
+	case cluster.OperationPhaseKubernetes:
+		return KubernetesPhaseSchemaPath, nil
+
+	case cluster.OperationPhaseDistribution:
+		return DistributionPhaseSchemaPath, nil
+
+	case cluster.OperationPhasePlugins:
+		return PluginsPhaseSchemaPath, nil
+
+	case cluster.OperationPhaseAll:
+		return AllPhaseSchemaPath, nil
+
+	default:
+		return "", fmt.Errorf("%w: %s", ErrUnsupportedPhase, phase)
+	}
 }
